add separate examples for device light and geolocation
