+++ examples/device-light/index.html
@@ -1,0 +1,7 @@
+<html>
+<body>
+  <div id="app"></div>
+  <script src="http://fb.me/react-with-addons-0.12.0.js"></script>
+  <script src="./js/index.js"></script>
+</body>
+</html>
